requirements_test.txt
black==22.3.0
codespell==2.0.0
flake8==3.9.2
mypy==0.901
pydocstyle==6.1.1
pylint==2.8.2
pylint-strict-informational==0.1
pytest-homeassistant-custom-component==0.13.202
